+++ app/helpers/dataset_users_helper.rb
@@ -1,0 +1,2 @@
+module DatasetUsersHelper
+end