+++ app/models/dataset_user.rb
@@ -1,0 +1,7 @@
+class DatasetUser < ActiveRecord::Base
+
+  # Model Relationships
+  belongs_to :dataset
+  belongs_to :user
+
+end