+++ db/migrate/20131101182317_create_dataset_users.rb
@@ -1,0 +1,12 @@
+class CreateDatasetUsers < ActiveRecord::Migration
+  def change
+    create_table :dataset_users do |t|
+      t.integer :dataset_id
+      t.integer :user_id
+      t.boolean :editor, null: false, default: false
+      t.boolean :approved
+
+      t.timestamps
+    end
+  end
+end