+++ db/migrate/20131101183940_add_public_files_to_datasets.rb
@@ -1,0 +1,5 @@
+class AddPublicFilesToDatasets < ActiveRecord::Migration
+  def change
+    add_column :datasets, :public_files, :boolean, null: false, default: true
+  end
+end