+++ test/helpers/dataset_users_helper_test.rb
@@ -1,0 +1,4 @@
+require 'test_helper'
+
+class DatasetUsersHelperTest < ActionView::TestCase
+end